requirements.txt
docx2pdf==0.1.8
fastapi==0.116.1
fpdf==1.7.2
pandas==2.3.1
protobuf==6.32.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.43.2
streamlit_extras==0.7.6
streamlit_lottie==0.0.5
uvicorn==0.35.0
